XIVY-11599 Cleanup start event beans, create IntermediateEvent
