+++ engine-cockpit/src/ch/ivyteam/enginecockpit/monitor/StartEventBean.java
@@ -2,20 +2,14 @@ package ch.ivyteam.enginecockpit.monitor;
 
 import java.lang.management.ManagementFactory;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.ViewScoped;
-import javax.management.AttributeNotFoundException;
-import javax.management.InstanceNotFoundException;
-import javax.management.MBeanException;
 import javax.management.MalformedObjectNameException;
 import javax.management.ObjectName;
-import javax.management.ReflectionException;
 
-import ch.ivyteam.enginecockpit.monitor.unit.Unit;
-import ch.ivyteam.enginecockpit.util.DateUtil;
+import ch.ivyteam.enginecockpit.monitor.events.start.StartEvent;
 import ch.ivyteam.enginecockpit.util.ErrorHandler;
 import ch.ivyteam.log.Logger;
 
@@ -25,10 +19,10 @@ public class StartEventBean {
   private static final Logger LOGGER = Logger.getPackageLogger(StartEventBean.class);
   private static final ErrorHandler HANDLER = new ErrorHandler("msgs", LOGGER);
 
-  private List<Bean> beans;
-  private List<Bean> filteredBeans;
+  private List<StartEvent> beans;
+  private List<StartEvent> filteredBeans;
   private String filter;
-  private Bean selected;
+  private StartEvent selected;
 
   public StartEventBean() {
     refresh();
@@ -39,7 +33,7 @@ public class StartEventBean {
       beans = ManagementFactory.getPlatformMBeanServer()
               .queryNames(new ObjectName("ivy Engine:type=Process Start Event Bean,application=*,pm=*,pmv=*,name=*"), null)
               .stream()
-              .map(Bean::new)
+              .map(StartEvent::new)
               .toList();
       beans = new ArrayList<>(beans);
     } catch (MalformedObjectNameException ex) {
@@ -47,15 +41,15 @@ public class StartEventBean {
     }
   }
 
-  public List<Bean> getBeans() {
+  public List<StartEvent> getBeans() {
     return beans;
   }
 
-  public List<Bean> getFilteredBeans() {
+  public List<StartEvent> getFilteredBeans() {
     return filteredBeans;
   }
 
-  public void setFilteredBeans(List<Bean> filteredBeans) {
+  public void setFilteredBeans(List<StartEvent> filteredBeans) {
     this.filteredBeans = filteredBeans;
   }
 
@@ -67,158 +61,11 @@ public class StartEventBean {
     this.filter = filter;
   }
 
-  public Bean getSelected() {
+  public StartEvent getSelected() {
     return selected;
   }
 
-  public void setSelected(Bean selected) {
+  public void setSelected(StartEvent selected) {
     this.selected = selected;
-  }
-
-  public static final class Bean {
-
-    private static final String NOT_AVAILABLE = "n.a.";
-    private static final Object[] EMPTY_PARAMS = new Object[0];
-    private static final String[] EMPTY_TYPES = new String[0];
-    private ObjectName name;
-
-    private Bean(ObjectName name) {
-      this.name = name;
-    }
-
-    public String getName() {
-      return name.getKeyProperty("name");
-    }
-
-    public String getBeanName() {
-      return readStringAttribute("name");
-    }
-
-    public String getBeanDescription() {
-      return readStringAttribute("description");
-    }
-
-    public String getFullRequestPath() {
-      return getApplication() +
-             "/"+
-             getPm() +
-             "$" +
-             getPmv() +
-             "/" +
-             readStringAttribute("requestPath");
-    }
-
-    public String getApplication() {
-      return name.getKeyProperty("application");
-    }
-
-    public String getPm() {
-      return name.getKeyProperty("pm");
-    }
-
-    public String getPmv() {
-      return name.getKeyProperty("pmv");
-    }
-
-    public String getRequestPath() {
-      return readStringAttribute("requestPath");
-    }
-
-    public boolean isRunning() {
-      return (Boolean)readAttribute("running");
-    }
-
-    public String getNextPollTime() {
-      return getDateAttribute("nextPollTime");
-    }
-
-    public long getTimeUntilNextPoll() {
-      return readLongAttribute("timeUntilNextPoll");
-    }
-
-    public String getTimeUntilNextPollFormated() {
-      return formatMillis(readLongAttribute("timeUntilNextPoll"));
-    }
-
-    public long getExecutions() {
-      return readLongAttribute("processExecutions");
-    }
-
-    public long getErrors() {
-      return readLongAttribute("processExecutionErrors");
-    }
-
-    public void poll() {
-      try {
-        ManagementFactory.getPlatformMBeanServer().invoke(name, "pollNow", EMPTY_PARAMS, EMPTY_TYPES);
-      } catch (InstanceNotFoundException | ReflectionException | MBeanException ex) {
-        HANDLER.showError("Cannot poll bean", ex);
-      }
-    }
-
-    public void start() {
-      try {
-        ManagementFactory.getPlatformMBeanServer().invoke(name, "start", EMPTY_PARAMS, EMPTY_TYPES);
-      } catch (InstanceNotFoundException | ReflectionException | MBeanException ex) {
-        HANDLER.showError("Cannot start bean", ex);
-      }
-    }
-
-    public void stop() {
-      try {
-        ManagementFactory.getPlatformMBeanServer().invoke(name, "stop", EMPTY_PARAMS, EMPTY_TYPES);
-      } catch (InstanceNotFoundException | ReflectionException | MBeanException ex) {
-        HANDLER.showError("Cannot stop bean", ex);
-      }
-    }
-
-    private String formatMillis(long value) {
-      return format(value, Unit.MILLI_SECONDS);
-    }
-
-    private String format(long value, Unit baseUnit) {
-      if (value < 0) {
-        return "n.a.";
-      }
-      Unit unit = baseUnit;
-      var scaledValue = baseUnit.convertTo(value, unit);
-      while (shouldScaleUp(unit, scaledValue)) {
-        unit = unit.scaleUp();
-        scaledValue = baseUnit.convertTo(value, unit);
-      }
-      return scaledValue+" "+ unit.symbol();
-    }
-
-    private boolean shouldScaleUp(Unit unit, long scaledValue) {
-      if (Unit.MICRO_SECONDS.equals(unit) || Unit.MILLI_SECONDS.equals(unit)) {
-        return scaledValue >= 1000;
-      }
-      return scaledValue >= 100;
-    }
-
-    private String getDateAttribute(String attributeName) {
-      Date date = (Date)readAttribute(attributeName);
-      if (date == null) {
-        return NOT_AVAILABLE;
-      }
-      return DateUtil.formatDate(date);
-    }
-
-    private long readLongAttribute(String attribute) {
-      return (long)readAttribute(attribute);
-    }
-
-    private String readStringAttribute(String attribute) {
-      return (String)readAttribute(attribute);
-    }
-
-    Object readAttribute(String attribute) {
-      try {
-        return ManagementFactory.getPlatformMBeanServer().getAttribute(name, attribute);
-      } catch (InstanceNotFoundException | AttributeNotFoundException | ReflectionException | MBeanException ex) {
-        HANDLER.showError("Cannot read attribute " + attribute, ex);
-        return "";
-      }
-    }
   }
 }
