+++ engine-cockpit/webContent/includes/components/monitor/startevent-firings.xhtml
@@ -16,16 +16,16 @@
     <p:panelGrid columns="6" columnClasses="p-col-12 p-md-12 p-lg-1, p-col-12- p-md-12 p-lg-3, p-col-12- p-md-1 p-lg-1, p-col-12- p-md-1 p-lg-3, p-col-12- p-md-12 p-lg-1, p-col-12- p-md-12 p-lg-3" layout="flex"
       styleClass="grey-label-panel ui-fluid">
       <p:outputLabel for="executions" value="Executions" />
-      <h:outputText id="executions" value="#{startEventDetailBean.executions}"/>
+      <h:outputText id="executions" value="#{startEventDetailBean.event.executions}"/>
       <p:outputLabel for="duration" value="Execution Duration" />
       <h:panelGroup>
-        <h:outputText id="duration" value="#{startEventDetailBean.minExecutionTime} / #{startEventDetailBean.avgExecutionTime} / #{startEventDetailBean.maxExecutionTime}"/>
+        <h:outputText id="duration" value="#{startEventDetailBean.event.minExecutionTime} / #{startEventDetailBean.event.avgExecutionTime} / #{startEventDetailBean.event.maxExecutionTime}"/>
         <p:tooltip for="duration" value="min / avg / max"/>
       </h:panelGroup>
       <p:outputLabel for="errors" value="Errors" />
-      <h:outputText id="errors" value="#{startEventDetailBean.errors}"/>
+      <h:outputText id="errors" value="#{startEventDetailBean.event.errors}"/>
     </p:panelGrid>
-    <p:dataTable value="#{startEventDetailBean.firings}" var="firing" widgetVar="firingTable" styleClass="ui-fluid"
+    <p:dataTable value="#{startEventDetailBean.event.firings}" var="firing" widgetVar="firingTable" styleClass="ui-fluid"
             paginator="true" rows="20" paginatorPosition="bottom" paginatorAlwaysVisible="false"
             id="firingTable" style="min-width: 950px;">
       <p:column headerText="Timestamp" sortBy="#{firing.timestamp}" filterBy="#{firing.timestamp}">
