+++ engine-cockpit/webContent/includes/components/monitor/startevent-poll.xhtml
@@ -18,30 +18,30 @@
     <p:panelGrid columns="2" columnClasses="p-col-12 p-md-3 p-lg-3, p-col-12 p-md-9 p-lg-9" layout="flex"
       styleClass="grey-label-panel ui-fluid">
       <p:outputLabel for="polls" value="Polls" />
-      <h:outputText id="polls" value="#{startEventDetailBean.polls}"/>
+      <h:outputText id="polls" value="#{startEventDetailBean.event.polls}"/>
       <p:outputLabel for="pollDuration" value="Poll Duration" />
       <h:panelGroup>
-        <h:outputText id="pollDuration" value="#{startEventDetailBean.minPollTime} / #{startEventDetailBean.avgPollTime} / #{startEventDetailBean.maxPollTime}"/>
+        <h:outputText id="pollDuration" value="#{startEventDetailBean.event.minPollTime} / #{startEventDetailBean.event.avgPollTime} / #{startEventDetailBean.event.maxPollTime}"/>
         <p:tooltip for="pollDuration" value="min / avg / max"/>
       </h:panelGroup>
       <p:outputLabel for="pollErrors" value="Poll Errors" />
-      <h:outputText id="pollErrors" value="#{startEventDetailBean.pollErrors}"/>
+      <h:outputText id="pollErrors" value="#{startEventDetailBean.event.pollErrors}"/>
       <p:outputLabel for="lastPollError" value="Last Poll Error" />
-      <cc:Error id="lastPollError" error="#{startEventDetailBean.lastPollError}"/>
+      <cc:Error id="lastPollError" error="#{startEventDetailBean.event.lastPollError}"/>
       <p:outputLabel for="pollConfiguration" value="Poll Configuration" />
-      <h:outputText id="pollConfiguration" value="#{startEventDetailBean.humanReadablePollConfiguration} (#{startEventDetailBean.pollConfiguration})"/>
+      <h:outputText id="pollConfiguration" value="#{startEventDetailBean.event.humanReadablePollConfiguration} (#{startEventDetailBean.event.pollConfiguration})"/>
       <p:outputLabel for="nextPoll" value="Next Poll" />
-      <h:outputText id="nextPoll" value="#{startEventDetailBean.timeUntilNextPollFormated} (#{startEventDetailBean.nextPollTime})"/>
+      <h:outputText id="nextPoll" value="#{startEventDetailBean.event.timeUntilNextPollFormated} (#{startEventDetailBean.event.nextPollTime})"/>
     </p:panelGrid>
   </h:form>
-  <p:dialog style="min-width: 300px;" header="Poll Bean#{startEventDetailBean.name}" id="pollBeanDialog"
+  <p:dialog style="min-width: 300px;" header="Poll Bean#{startEventDetailBean.event.name}" id="pollBeanDialog"
       widgetVar="pollBeanDialog" responsive="true" closeOnEscape="true">
     <h:form id="pollBean" styleClass="custom-dialog-form">
       <p:staticMessage severity="warn" detail="Scheduling to poll a bean can have negative side effects like." />
       <div class="custom-dialog-footer">
         <p:commandButton id="cancel" type="button" onclick="PF('pollBeanDialog').hide();" value="Cancel" styleClass="ui-button-secondary ui-button-flat modal-input-button"/>
         <p:commandButton id="poll" onclick="PF('pollBeanDialog').hide();" value="Poll" icon="si si-controls-play"
-            styleClass="modal-input-button" actionListener="#{startEventDetailBean.poll()}" update="startEventPoll"/>
+            styleClass="modal-input-button" actionListener="#{startEventDetailBean.event.poll()}" update="startEventPoll"/>
       </div>
     </h:form>
    </p:dialog>
